+++ src/ai/backend/manager/config.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+import json
+
 """
 Configuration Schema on etcd
 ----------------------------
@@ -208,6 +210,7 @@ from ai.backend.common.identity import get_instance_id
 from ai.backend.common.logging import BraceStyleAdapter
 from ai.backend.common.types import (
     HostPortPair,
+    RoundRobinState,
     SlotName,
     SlotTypes,
     current_resource_slots,
@@ -490,6 +493,7 @@ shared_config_iv = t.Dict(
                 ): session_hang_tolerance_iv,
             },
         ).allow_extra("*"),
+        t.Key("roundrobin_states", default=None): t.Null | tx.RoundRobinStatesJSONString,
     }
 ).allow_extra("*")
 
@@ -869,3 +873,38 @@ class SharedConfig(AbstractConfig):
             self.data["redis"]["addr"][1]
         ).with_password(self.data["redis"]["password"]) / str(db)
         return url
+
+    async def get_roundrobin_state(
+        self, resource_group_name: str, architecture: str
+    ) -> RoundRobinState | None:
+        """
+        Return the roundrobin state for the given resource group and architecture.
+        If given resource group's roundrobin states or roundrobin state of the given architecture is not found, return None.
+        """
+        if (rr_state_str := await self.get_raw("roundrobin_states")) is not None:
+            rr_states_dict: dict[str, dict[str, Any]] = json.loads(rr_state_str)
+            resource_group_rr_states_dict = rr_states_dict.get(resource_group_name, None)
+
+            if resource_group_rr_states_dict is not None:
+                rr_state_dict = resource_group_rr_states_dict.get(architecture, None)
+
+                if rr_state_dict is not None:
+                    return RoundRobinState(
+                        schedulable_group_id=rr_state_dict["schedulable_group_id"],
+                        next_index=rr_state_dict["next_index"],
+                    )
+
+        return None
+
+    async def put_roundrobin_state(
+        self, resource_group_name: str, architecture: str, state: RoundRobinState
+    ) -> None:
+        """
+        Update the roundrobin states using the given resource group and architecture key.
+        """
+        rr_states_dict = json.loads(await self.get_raw("roundrobin_states") or "{}")
+        if resource_group_name not in rr_states_dict:
+            rr_states_dict[resource_group_name] = {}
+
+        rr_states_dict[resource_group_name][architecture] = state.to_json()
+        await self.etcd.put("roundrobin_states", json.dumps(rr_states_dict))