+++ src/ai/backend/storage/context.py
@@ -39,6 +39,7 @@ from .plugin import (
 )
 from .purestorage import FlashBladeVolume
 from .types import VolumeInfo
+from .vast import VASTVolume
 from .vfs import BaseVolume
 from .weka import WekaVolume
 from .xfs import XfsVolume
@@ -59,6 +60,7 @@ DEFAULT_BACKENDS: Mapping[str, Type[AbstractVolume]] = {
     GPFSVolume.name: GPFSVolume,  # IBM SpectrumScale or GPFS
     "spectrumscale": GPFSVolume,  # IBM SpectrumScale or GPFS
     CephFSVolume.name: CephFSVolume,
+    VASTVolume.name: VASTVolume,
 }